requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==1.4.54
psycopg2-binary==2.9.10
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
requests==2.32.5
beautifulsoup4==4.13.4
openai==1.102.0
google-generativeai==0.8.5
apache-airflow==2.11.0
python-dotenv==1.1.1
